fix(sessions) : bug fix about ghost window
1. Problem

Sometimes, although neovim had only one window and it is focused, there
are two `windows` data in session_data file.
One is focused window, the other is about scratch file.
It makes error when I load session because there are no valid buffer in
one window.

I don't know what is the reason. It invokes not every time.

2. Solution

1) utils.filter_ignored() function can check the buffer is valid now.
It can makes the condition which file_data is inserted in buffers table is
removed.

2) Not only inserting buffer situation, but also inserting window
   situation, check whether the buffer in window is valid