fix(sessions) : fix issue about detecting cursor position of unloaded buffer
1. Problem

Before fix, cursor locations of listed buffers is saved by getbufinfo()
only. but If you quit neovim without view the buffer which is listed but
unloaded, the result of getbufinfo() will be broken. line number is
gone.
It makes error to move cursor when I move focus to the unloaded buffer after session is
loaded.

2. Solution

Create local variable `loaded_session_data` to save loaded session file data.
It would be used when session is saved, if some file are not loaded and
there are no information about line number of cursor in these one, don't
update line number information of the buffer.